requirements.txt
flet==0.28.3
psutil==7.2.2
Pillow==12.1.0
pywin32==311
pystray==0.19.5
